schedule_redundant_bm_write does a lot of seeking back and forth between data and the 512 bitmap
This was meant to be an optimization to keep I/O sequential (see comment in source code),
except in this case it is failing badly at keeping I/O sequential: it will itself introduce a lot of non-sequential I/O.

TBC: does this have a perf throughput reduction on raw devices as the comment claimed?
Albeit nowadays we are more interested in optimizing throughput on fast storage.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>